+++ data/exposed/src/main/kotlin/io/bluetape4k/exposed/dao/StringEntity.kt
@@ -19,13 +19,14 @@ abstract class StringEntity(id: EntityID<String>): Entity<String>(id)
  * @param [entityType] The expected [StringEntity] type. This can be left `null` if it is the class of type
  * argument [E] provided to this [StringEntityClass] instance. If this `StringEntityClass` is defined as a companion
  * object of a custom `StringEntity` class, the parameter will be set to this immediately enclosing class by default.
- * @sample org.jetbrains.exposed.sql.tests.shared.DDLTests.testDropTableFlushesCache
+ *
+ * @sample io.bluetape4k.exposed.dao.TimebasedUUIDBase62TableTest.E1
+ * 
  * @param [entityCtor] The function invoked to instantiate a [StringEntity] using a provided [EntityID] value.
  * If a reference to a specific constructor or a custom function is not passed as an argument, reflection will
  * be used to determine the primary constructor of the associated entity class on first access. If this `StringEntityClass`
  * is defined as a companion object of a custom `StringEntity` class, the constructor will be set to that of the
  * immediately enclosing class by default.
- * @sample org.jetbrains.exposed.sql.tests.shared.entities.EntityTests.testExplicitEntityConstructor
  */
 abstract class StringEntityClass<out E: StringEntity>(
     table: IdTable<String>,
